UseSpawnForTrainSubprocess: Move model spawn wrapping to LocalModelTrainer
This is the only place we know the start method and it's the only place we
know that we need this safe wrapping

https://github.com/caikit/caikit/issues/384
